Fix Vercel deployment failed
