+++ package.json
@@ -40,5 +40,8 @@
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
+  },
+  "resolutions": {
+    "typeorm": "0.2.24"
   }
 }
